% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{drought}
\alias{drought}
\title{Compute drought indices based on daily climate}
\usage{
drought(
  data,
  data_GS,
  threshold_growth = 4,
  threshold_prec = 0,
  by_year = TRUE
)
}
\arguments{
\item{data}{Data frame created from the generateWeather() function of BioSIM generated using the "Climatic_Daily" model}

\item{data_GS}{data.frame produced by \link{growing_season}}

\item{threshold_growth}{Temperature threshold (Celsius degrees) for calculating growing degree-days during the drought event, which are computed based on this threshold.}

\item{threshold_prec}{Daily precipitation (mm) threshold for inducing a drought event. A drought event will start when the daily precipitation is below this threshold, and will continue until the daily precipitation is higher than this threshold.}

\item{by_year}{boolean. Summarise by year. If false, a daily output is generated.}
}
\description{
This function identifies drought events occurring during the growing season and computes drought indices based on a precipitation threshold. The drought event starts whenever the daily precipitation is below that specified by the \code{threshold_prec} argument and ends when the daily precipitation is above this threshold.
The function cumulates degree-days (above the threshold indicated by \code{threshold_growth}) during the drought events. When \code{by_year = TRUE}, the function returns a data frame containing the maximum GDD for a single event (column \code{droughtevent_dd}) and the total amount of GDD cumulated during all the drought events that occurred in the entire growing season  (column \code{droughtGS_dd}).
The function also returns the maximum number of days for a single drought event \code{droughtevent_days}) and the total number of drought days over the entire growing season \code{droughtGS_days}).
When \code{by_year = FALSE}, the function returns a data frame containing daily values. This may be used to characterize each drought event.
}
\examples{
file = system.file("extdata", "dayclim_GS.txt", package="ExtremeEvents")
data <- read.table(file)

GS = growing_season(data)
GS

data = drought(data, GS)
data

}
