% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{growing_season}
\alias{growing_season}
\alias{drought}
\title{Calcul de la saison de croissance et de la sécheresse à partir des données BioSIM}
\usage{
growing_season(
  data,
  min_growing_day = 115,
  max_growing_day = 227,
  threshold_growth = 4,
  threshold_dormancy = 0
)

drought(
  data,
  data_GS,
  threshold_growth = 4,
  threshold_prec = 0,
  by_year = TRUE
)
}
\arguments{
\item{data}{data.frame BioSIM data}

\item{min_growing_day}{jour julien, Default 15 avril}

\item{max_growing_day}{jour julien, Default 15 août}

\item{threshold_growth}{seuil de température pour la croissance}

\item{threshold_dormancy}{seuil de température pour la dormance}

\item{data_GS}{data.frame produit par \link{growing_season}}

\item{threshold_prec}{seuil de précipitation journalière pour les sécheresses}

\item{by_year}{boolean. Summarise by year}
}
\description{
Calcul de la saison de croissance et de la sécheresse à partir des données BioSIM.
growing_season_and_drought_yearly est la fonction principale qui fait tout.
}
\examples{
file = system.file("extdata", "dayclim_GS.txt", package="ExtremeEvents")
data <- read.table(file)

GS = growing_season(data)
GS

data = drought(data, GS)
data
}
