% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{lsf}
\alias{lsf}
\title{Compute late spring frost indices based on daily climate}
\usage{
lsf(
  data,
  data_GS,
  threshold_frost = -4,
  threshold_growth = 4,
  threshold_lsf = 196,
  by_year = TRUE
)
}
\arguments{
\item{data}{Data frame created from the generateWeather() function of BioSIM generated using the "Climatic_Daily" model}

\item{data_GS}{data.frame produced by \link{growing_season}}

\item{threshold_frost}{Daily precipitation (mm) threshold for inducing a drought event. A drought event will start when the daily precipitation is below this threshold, and will continue until the daily precipitation is higher than this threshold.}

\item{threshold_growth}{Temperature threshold (Celsius degrees) for cumulating growing degree-days before the frost event.}

\item{threshold_lsf}{Maximum date (day of year format) for considering late spring frost. This prevent identifying the end of the growing season as a LSF.}

\item{by_year}{boolean. Summarise by year. If false, a daily output is generated.}
}
\description{
This function identifies late spring frost (LSF) events occurring during the growing season. It computes a LSF severity index, which we defined as the number of growing degree-days cumulated before the frost events.
The temperature threshold (Celsius degrees) for identifying a LSF event is defined by the \code{threshold_frost} argument. This threshold should represent a temperature below which damage to the tree are expected.
The growing degree-days cumulate from the start of the growing season until a LSF occurs. The temperature above which to cumulate growing degree-days can be specified by the \code{threshold_growth} argument.
When \code{by_year = TRUE}, the function returns a data frame containing the maximum GDD for a single event (column \code{lsfevent_dd}) and the total amount of GDD cumulated before the last frost event of the growing season (column \code{lsfGS_dd}).
Similarly, the function returns a columns for the number of days with mean temperature abov the growth threshold before a the worst LSF event (column \code{lsfevent_days}, and the total number of days before the last LSF event of the growing season (column \code{lsfGS_days}).
When \code{by_year = FALSE}, the function returns as data frame containing daily values of \code{lsfevent_dd} and \code{lsfGS_dd}. This may be used to characterize each LSF event.
}
